refactor: improve UI styling and extract game logic to custom hook
- Remove unused levels page component
- Extract game logic from Game component to custom hook
- Update styling for cards, navbar, and game screens
- Add responsive design and animations
- Move global styles to App.css
