+++ Card-game/src/Components/gameComponents/cardList/cardsList.css
@@ -1,20 +1,34 @@
 .wrapper {
   display: grid;
-  gap: 10px;
-  padding: 10px;
-  border: 1px solid black;
-  width: 600px;
-  min-height: 60vh;
+  gap: 1.2rem;
+  justify-content: center;
+  align-items: center;
+  margin: 2rem auto;
+  padding: 2rem;
+  border-radius: 1rem;
+  background: rgba(255, 255, 255, 0.02);
+  backdrop-filter: blur(12px);
+  box-shadow: 0 0 20px rgba(6, 182, 212, 0.1);
+  transition: all 0.3s ease-in-out;
 }
-.wrapper.list-2 {
-  grid-template-columns: repeat(2, 1fr);
-  grid-template-rows: repeat(2, 1fr);
+
+/* Level-specific grid layout */
+.list-2 {
+  grid-template-columns: repeat(2, 120px);
 }
-.wrapper.list-4 {
-  grid-template-columns: repeat(4, 1fr);
-  grid-template-rows: repeat(4, 1fr);
+.list-4 {
+  grid-template-columns: repeat(4, 100px);
 }
-.wrapper.list-6 {
-  grid-template-columns: repeat(6, 1fr);
-  grid-template-rows: repeat(6, 1fr);
+.list-6 {
+  grid-template-columns: repeat(6, 80px);
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+  .list-4 {
+    grid-template-columns: repeat(2, 100px);
+  }
+  .list-6 {
+    grid-template-columns: repeat(3, 80px);
+  }
 }
