+++ Card-game/src/Components/gameComponents/carditem/card.css
@@ -1,41 +1,54 @@
 .Card {
   aspect-ratio: 1;
-  border: 1px solid black;
-  border-radius: 10px;
+  width: 100%;
+  border-radius: 12px;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
+  cursor: pointer;
   display: flex;
   justify-content: center;
   align-items: center;
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: contain;
-  cursor: pointer;
+  transition: transform 0.25s ease, box-shadow 0.25s ease;
+  background-color: #1e293b;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  overflow: hidden;
+  backface-visibility: hidden;
+  perspective: 1000px;
 }
+
 .Card:hover {
-  border: 2px solid black;
+  transform: scale(1.05);
+  box-shadow: 0 6px 20px rgba(6, 182, 212, 0.4); /* cyan glow */
+  border: 1px solid rgba(6, 182, 212, 0.5);
 }
+
+/* Flip Animation */
 .flipped {
-  webkit-animation: flip-horizontal-bottom 0.4s
-    cubic-bezier(0.455, 0.03, 0.515, 0.955) both;
   animation: flip-horizontal-bottom 0.4s cubic-bezier(0.455, 0.03, 0.515, 0.955)
     both;
 }
-@-webkit-keyframes flip-horizontal-bottom {
+
+@keyframes flip-horizontal-bottom {
   0% {
-    -webkit-transform: rotateX(-180deg);
     transform: rotateX(-180deg);
   }
   100% {
-    -webkit-transform: rotateX(0);
     transform: rotateX(0);
   }
 }
-@keyframes flip-horizontal-bottom {
+.matched {
+  border: 2px solid #06b6d4;
+  box-shadow: 0 0 15px rgba(6, 182, 212, 0.6);
+  animation: pulse 0.6s ease-in-out;
+}
+
+@keyframes pulse {
   0% {
-    -webkit-transform: rotateX(-180deg);
-    transform: rotateX(-180deg);
+    transform: scale(1.05);
   }
   100% {
-    -webkit-transform: rotateX(0);
-    transform: rotateX(0);
+    transform: scale(1);
   }
 }
