+++ Card-game/src/Components/gameComponents/navBar/navbar.css
@@ -1,13 +1,28 @@
 .Bar {
-  border: 1px solid black;
-  position: relative;
-  min-width:90%;
-}
-.Bar > .nav {
+  background: rgba(255, 255, 255, 0.05);
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  margin: 1rem auto;
+  padding: 1rem 2rem;
+  border-radius: 0.75rem;
+  width: fit-content;
   display: flex;
+  justify-content: center;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.2);
+}
+
+.nav {
   list-style: none;
-  flex-direction: row;
-  gap: 5px;
-  justify-content: space-around;
-  width: 100%;
+  display: flex;
+  gap: 2rem;
+  margin: 0;
+  padding: 0;
+  color: #f1f5f9;
+  font-weight: 500;
+  font-size: 1rem;
+}
+
+.nav li {
+  display: flex;
+  align-items: center;
 }
