--- Card-game/src/pages/levels.tsx
@@ -1,9 +1,0 @@
-import './_pages.css'
-
-function Levels() {
-  return (
-    <div className='screen'>Levels</div>
-  )
-}
-
-export default Levels
